Refact-Join 회원가입 로직을 컴포넌트 분리시키려 했으나, 시간이 너무 많이 빨려서 1차적인 분리까지만 진행, PersonalInfoModal 부분 업데이트 필요
