+++ src/shared/Router.tsx
@@ -12,13 +12,7 @@ import Statistics from '../pages/statistics/Statistics';
 import OrderMenu from '../pages/orderpage/OrderMenu';
 import OrderCustom from '../pages/orderpage/OrderCustom';
 import Mobile from './layout/Mobile';
-import { supabase } from '../supabase';
 import { PrivateRoute } from '../pages/statistics/PrivateRoute';
-
-console.log('Router', supabase);
-// interface ProfileProps {
-//   session: supabase.auth.Session | null; // Use the Session type from supabase
-// }
 
 const Router = () => {
   return (
